requirements.txt
pyyaml
twitter
ipdb
nltk
Flask
flask-bootstrap
Jinja2
Werkzeug
